Set mesh hold-out tools moved under Display tools. Removed old set hold-out tools.
