+++ share/config/menu.json
@@ -88,6 +88,11 @@
             "display_tools/reset_object_colour",
             "display_tools/toggle_object_motion_trail",
             "display_tools/create_screen_space_motion_trail",
+            "display_tools/---Mesh Hold-Out",
+            "display_tools/enable_holdout_all_meshes",
+            "display_tools/disable_holdout_all_meshes",
+            "display_tools/enable_holdout_selected_meshes",
+            "display_tools/disable_holdout_selected_meshes",
             "zdepth_tools/---Z-Depth Screen-Space Tools",
             "zdepth_tools/screen_z_manipulator",
             "zdepth_tools/screen_space_rig_bake",
@@ -98,10 +103,6 @@
             "mesh_tools/---Create",
             "mesh_tools/mesh_from_locators",
             "mesh_tools/---Hold-Outs",
-            "mesh_tools/enable_holdouts_for_all_meshes",
-            "mesh_tools/disable_holdouts_for_all_meshes",
-            "mesh_tools/enable_holdouts_for_selected_meshes",
-            "mesh_tools/disable_holdouts_for_selected_meshes",
             "general_tools/---Parenting",
             "general_tools/reparent_under_node2",
             "general_tools/unparent_to_world2",
