+++ share/config/shelf.json
@@ -85,6 +85,11 @@
             "display_tools/display_popup/reset_object_colour",
             "display_tools/display_popup/toggle_object_motion_trail",
             "display_tools/display_popup/create_screen_space_motion_trail",
+            "display_tools/display_popup/---Mesh Hold-Out",
+            "display_tools/display_popup/enable_holdout_all_meshes",
+            "display_tools/display_popup/disable_holdout_all_meshes",
+            "display_tools/display_popup/enable_holdout_selected_meshes",
+            "display_tools/display_popup/disable_holdout_selected_meshes",
             "zdepth_tools/zdepth_popup/---Z-Depth Screen-Space Tools",
             "zdepth_tools/zdepth_popup/screen_z_manipulator",
             "zdepth_tools/zdepth_popup/screen_space_rig_bake",
@@ -94,11 +99,6 @@
             "zdepth_tools/zdepth_popup/screen_z_transform_bake",
             "mesh_tools/mesh_tools_popup/---Create",
             "mesh_tools/mesh_tools_popup/mesh_from_locators",
-            "mesh_tools/mesh_tools_popup/---Hold-Outs",
-            "mesh_tools/mesh_tools_popup/enable_holdouts_for_all_meshes",
-            "mesh_tools/mesh_tools_popup/disable_holdouts_for_all_meshes",
-            "mesh_tools/mesh_tools_popup/enable_holdouts_for_selected_meshes",
-            "mesh_tools/mesh_tools_popup/disable_holdouts_for_selected_meshes",
             "general_tools/gen_tools_popup/---Parenting",
             "general_tools/gen_tools_popup/reparent_under_node2",
             "general_tools/gen_tools_popup/unparent_to_world2",
